Update Dockerfile to compile GCC with Ada
Rather than using apt-get to get GNAT which has an uncontrolled
version instead compile GCC with Ada enabled to get a known
version of the compiler. This is based on the docker-library/gcc
Docker images. The next steps for this would be to just create
a pull request putting this into docker-library/gcc image itself
as well as set up it up with Docker templates so you can make it
for a variety of targets.